FPBA-2 Account Controller update

+++ src/main/java/com/jaqg/banking/dto/CreateAccountRequestDTO.java
@@ -1,8 +1,10 @@
 package com.jaqg.banking.dto;
 
+import java.math.BigDecimal;
+
 public record CreateAccountRequestDTO(
       Long customerId,
       String accountName,
-      Double openingBalance
+      BigDecimal openingBalance
 ) {
 }
